Reorganize pill component
